Moved every database-related file to a new package inside the database package called "local". This is to prepare for the online database implementation
