+++ app/src/main/java/com/example/finalproject/custom_views/UserAdapter.java
@@ -18,8 +18,8 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.finalproject.R;
 import com.example.finalproject.activities.UsersActivity;
-import com.example.finalproject.database.AppDatabase;
-import com.example.finalproject.database.entities.User;
+import com.example.finalproject.database.local.AppDatabase;
+import com.example.finalproject.database.local.entities.User;
 import com.example.finalproject.util.Result;
 import com.example.finalproject.util.UserFilter;
 import com.example.finalproject.util.Util;
