+++ app/src/main/java/com/example/finalproject/database/local/Converters.java
@@ -1,4 +1,4 @@
-package com.example.finalproject.database;
+package com.example.finalproject.database.local;
 
 import androidx.room.TypeConverter;
